simplify and move tests
